some cleanup/housekeeping
I noticed that the travis deploy setup was not properly picking up the cookbook name so I set it explicitly.

Added and updated some other misc stuff in the metadata to match reality (for example the LICENSE in repo was MIT but the attribute in metadata said all rights reserved), chef version support, issue url, source code url, etc.

Signed-off-by: Ben Abrams <me@benabrams.it>